PDFtoPPT/requirements.txt
PyMuPDF
Pillow
python-pptx
